requirements.txt
# LLM and Agent Framework
anthropic>=0.18.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
langgraph>=0.0.20

# Document Processing
pdfplumber>=0.10.0
pypdf>=3.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0

# Embeddings and Vector Store
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.0

# Web UI
streamlit>=1.28.0
gradio>=4.0.0

# Utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
